Rename to TrackerAnimationInfoPanel
