+++ app/src/main/java/com/duckduckgo/app/tabs/ui/TabSwitcherItem.kt
@@ -21,7 +21,7 @@ import com.duckduckgo.app.tabs.model.TabEntity
 sealed class TabSwitcherItem(val id: String) {
 
     data class Tab(val tabEntity: TabEntity) : TabSwitcherItem(tabEntity.tabId)
-    data class TrackerAnimationTile(val trackerCount: Int) : TabSwitcherItem("TrackerAnimationTile") {
+    data class TrackerAnimationInfoPanel(val trackerCount: Int) : TabSwitcherItem("TrackerAnimationInfoPanel") {
 
         companion object {
             const val ANIMATED_TILE_NO_REPLACE_ALPHA = 0.4f
